add pagination and filter for admin and users

+++ controller/post/delete.php
@@ -8,15 +8,15 @@ include_once "../../model/blogs.php";
 use model\Database;
 use model\Blogs;
 
-if (isset($_GET['id']) && isset($_SESSION['user_id']) && isset($_SESSION['login'])){
+if (isset($_GET['id']) && isset($_SESSION['user']['user_id']) && isset($_SESSION['user']['login'])){
     $message = '';
     try {
-        if($_SESSION['login']) {
+        if($_SESSION['user']['login']) {
             $pdo = new Database();
             $pdo->setAttribute(PDO::ATTR_EMULATE_PREPARES, false);
             $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
             $post = Blogs::getById($pdo, $_GET['id']);
-            if($_SESSION['user_id'] == $post->getAuthorId()){
+            if($_SESSION['user']['user_id'] == $post->getAuthorId()){
                 $post->delete();
                 $message = 'Delete successfully!';
             }else{
@@ -27,7 +27,7 @@ if (isset($_GET['id']) && isset($_SESSION['user_id']) && isset($_SESSION['login'
         $message = 'Can not delete post cause: '.  $e->getMessage();
     }
 
-    $_SESSION['message'] = $message;
+    $_SESSION['user']['message'] = $message;
     header('Location: /blog/view/posts/list.php');
     exit;
 }
